refactor: restruct remain previous folder(#192)
