+++ src/fsd_shared/configs/constants/index.ts
@@ -2,3 +2,4 @@ export * from './messages';
 export * from './errorMessages';
 export * from './routes';
 export * from './notificationTabs';
+export * from './locker';
