+++ src/fsd_shared/configs/constants/locker.ts
@@ -1,0 +1,9 @@
+export const LOCKER_CONSTANT = () => {
+  const floor = {
+    SECOND: '2층',
+    THIRD: '3층',
+    FOURTH: '4층',
+  } as const;
+
+  return { floor };
+};
