+++ src/fsd_widgets/locker/ui/LockerListHeader.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 
-import { LOCKER_CONSTANT } from '@/utils';
+import { LOCKER_CONSTANT } from '@/fsd_shared';
 
 interface LockerListHeaderProps {
   lockerLocations: Locker.LockerLocationsResponseDto;
